refactor: . Move tests/*Group* to tests/Query/Group/
