+++ tests/Query/Group/TaskGroups.test.ts
@@ -2,19 +2,19 @@
  * @jest-environment jsdom
  */
 import moment from 'moment';
-import { FilenameField } from '../src/Query/Filter/FilenameField';
-import { Grouper, type GrouperFunction } from '../src/Query/Group/Grouper';
-import type { Task } from '../src/Task';
-import { PathField } from '../src/Query/Filter/PathField';
-import { TagsField } from '../src/Query/Filter/TagsField';
-import { FolderField } from '../src/Query/Filter/FolderField';
-import { TaskGroups } from '../src/Query/Group/TaskGroups';
-import { StatusTypeField } from '../src/Query/Filter/StatusTypeField';
-import { HappensDateField } from '../src/Query/Filter/HappensDateField';
-import { DueDateField } from '../src/Query/Filter/DueDateField';
-import { SearchInfo } from '../src/Query/SearchInfo';
-import { fromLine } from './TestHelpers';
-import { TaskBuilder } from './TestingTools/TaskBuilder';
+import { FilenameField } from '../../../src/Query/Filter/FilenameField';
+import { Grouper, type GrouperFunction } from '../../../src/Query/Group/Grouper';
+import type { Task } from '../../../src/Task';
+import { PathField } from '../../../src/Query/Filter/PathField';
+import { TagsField } from '../../../src/Query/Filter/TagsField';
+import { FolderField } from '../../../src/Query/Filter/FolderField';
+import { TaskGroups } from '../../../src/Query/Group/TaskGroups';
+import { StatusTypeField } from '../../../src/Query/Filter/StatusTypeField';
+import { HappensDateField } from '../../../src/Query/Filter/HappensDateField';
+import { DueDateField } from '../../../src/Query/Filter/DueDateField';
+import { SearchInfo } from '../../../src/Query/SearchInfo';
+import { fromLine } from '../../TestHelpers';
+import { TaskBuilder } from '../../TestingTools/TaskBuilder';
 
 window.moment = moment;
